init: Improve slightly and install vim plugins

+++ init.sh
@@ -1,18 +1,19 @@
 #!/bin/bash
+set -e
 
 dir=$PWD
 backup=$HOME/dotfiles.bak
 
-mkdir -p $backup
-
-for file in $dir/*
-do
+for file in $dir/*; do
     filename=$(basename $file)
-            echo $file
+    echo $file
     if [[ "$filename" != "$(basename $0)" ]]; then
+        if [ -e $HOME/.$filename ]; then
+            mkdir -p $backup
+            # Move existing dotfile to $backup
+            mv $HOME/.$filename $backup/
+	fi
         echo "Creating link for .$filename"
-        # Move existing dotfile to $backup
-        mv $HOME/.$filename $backup/
         # Create symlink
         ln -s $dir/$filename $HOME/.$filename
     fi
@@ -20,3 +21,6 @@ done
 
 # Pull in submodules
 git submodule init && git submodule update
+
+# Install vim plugins
+vim +'PlugInstall --sync' +qa
